Fetch data from firestore

+++ app/src/main/java/com/example/mapstemplate/activities/CreateItemActivity.kt
@@ -20,9 +20,9 @@ class CreateItemActivity : AppCompatActivity() {
             val name = findViewById<EditText>(R.id.item_title).text.toString()
             val address = findViewById<EditText>(R.id.item_location).text.toString()
             val description = findViewById<EditText>(R.id.item_description).text.toString()
-            var cost = 0f
+            var cost = 0.0
             try {
-                cost = findViewById<EditText>(R.id.item_cost).text.toString().toFloat()
+                cost = findViewById<EditText>(R.id.item_cost).text.toString().toDouble()
             } catch (e: Exception) {
                 print("ERROR : " + e.message)
             }
