+++ app/src/main/java/com/example/mapstemplate/itineraries/Itinerary.kt
@@ -3,8 +3,8 @@ package com.example.travelapp.itineraries
 class Itinerary(var name: String) : java.io.Serializable  {
     val steps: ArrayList<Step> = ArrayList()
 
-    fun calculateItineraryPrice(): Float {
-        var totalPrice = 0f
+    fun calculateItineraryPrice(): Double {
+        var totalPrice = 0.0
         for (step in steps) {
             totalPrice += step.price
         }
