+++ app/src/main/java/com/example/mapstemplate/itineraries/Step.kt
@@ -1,5 +1,5 @@
 package com.example.travelapp.itineraries
 
-class Step(var name: String, var address: String, var price: Float, var description: String) : java.io.Serializable  {
+class Step(var name: String, var address: String, var price: Double, var description: String) : java.io.Serializable  {
 
 }
